+++ packages/bunshine/src/HttpRouter/HttpRouter.spec.ts
@@ -286,7 +286,7 @@ describe('HttpRouter', () => {
     let server: Server;
     beforeEach(() => {
       app = new HttpRouter();
-      server = app.listen({ port: port++ });
+      server = app.listen({ port: 0 });
     });
     afterEach(() => {
       server.stop(true);