+++ packages/bunshine/src/middleware/serveFiles/serveFiles.spec.ts
@@ -18,7 +18,7 @@ describe('serveFiles middleware', () => {
   let server: Server;
   beforeEach(() => {
     app = new HttpRouter();
-    server = app.listen({ port: port++ });
+    server = app.listen({ port: 0 });
   });
   afterEach(() => {
     server.stop(true);