+++ packages/bunshine/src/responseFactories/file/file.spec.ts
@@ -12,14 +12,44 @@ describe('c.file()', () => {
   afterEach(() => {
     server.stop(true);
   });
-  it('should handle files with disposition="attachment"', async () => {
+  it('should handle paths with disposition="attachment"', async () => {
     app.get('/home.html', c =>
       c.file(`${import.meta.dir}/../../../testFixtures/home.html`, {
         disposition: 'attachment',
       })
     );
     const resp = await fetch(`${server.url}home.html`);
-    expect(resp).toBeInstanceOf(Response);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Disposition')).toBe(
+      'attachment; filename="home.html"'
+    );
+    const file = await resp.blob();
+    const text = await file.text();
+    expect(text).toBe('<h1>Welcome home</h1>\n');
+  });
+  it('should allow multiple headers with same name', async () => {
+    app.get('/home.html', c => {
+      return c.file(`${import.meta.dir}/../../../testFixtures/home.html`, {
+        headers: new Headers([
+          ['Hello', 'bun'],
+          ['Hello', 'world'],
+        ]),
+      });
+    });
+    const resp = await fetch(`${server.url}home.html`);
+    console.log('resp.headers', resp.headers);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Hello')).toBe('bun, world');
+  });
+  it('should handle BunFile with disposition="attachment"', async () => {
+    app.get('/home.html', c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/home.html`
+      );
+      return c.file(file, { disposition: 'attachment' });
+    });
+    const resp = await fetch(`${server.url}home.html`);
+    expect(resp.status).toBe(200);
     expect(resp.headers.get('Content-Disposition')).toBe(
       'attachment; filename="home.html"'
     );
@@ -198,6 +228,22 @@ describe('c.file()', () => {
       expect(resp.status).toBeOneOf([206, 416, 200]);
       expect(resp.headers.get('Content-Type')).toInclude(mime);
     });
+    it(`should detect mime from partial bytes Blob - ${name}`, async () => {
+      app.get(`/${name}`, async c => {
+        const file = Bun.file(
+          `${import.meta.dir}/../../../testFixtures/file-type/${name}`
+        );
+        const buffer = await file.bytes();
+        const blob = new Blob([buffer], { type: 'image/png' });
+        return c.file(blob);
+      });
+      const resp = await fetch(`${server.url}${name}`, {
+        headers: { Range: 'bytes=0-9' },
+      });
+      expect(resp.status).toBe(206);
+      expect(resp.headers.get('Content-Type')).toInclude(mime);
+      expect(resp.headers.get('Content-Length')).toBe('10');
+    });
     it(`should detect mime from bytes 0-1 request - ${name}`, async () => {
       app.get(`/${name}`, c =>
         c.file(`${import.meta.dir}/../../../testFixtures/file-type/${name}`)
@@ -208,5 +254,135 @@ describe('c.file()', () => {
       expect(resp.status).toBe(206);
       expect(resp.headers.get('Content-Type')).toInclude(mime);
     });
+    it(`should detect mime from HEAD request`, async () => {
+      app.headGet(`/${name}`, c =>
+        c.file(`${import.meta.dir}/../../../testFixtures/file-type/${name}`)
+      );
+      const resp = await fetch(`${server.url}${name}`, { method: 'HEAD' });
+      expect(resp.status).toBe(200);
+      expect(resp.headers.get('Content-Type')).toInclude(mime);
+    });
   }
+  it('should detect mime from ArrayBuffer', async () => {
+    app.get(`/video.mp4`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture-mp4v2.mp4.data`
+      );
+      const buffer = await file.arrayBuffer();
+      return c.file(buffer);
+    });
+    const resp = await fetch(`${server.url}video.mp4`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('video/mp4');
+  });
+  it('should detect mime from Bunfile', async () => {
+    app.get(`/my.pdf`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.pdf.data`
+      );
+      return c.file(file);
+    });
+    const resp = await fetch(`${server.url}my.pdf`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('application/pdf');
+  });
+  it('should detect mime from Uint8Array', async () => {
+    app.get(`/my.jpg`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.jpg.data`
+      );
+      const buffer = await file.bytes();
+      return c.file(buffer);
+    });
+    const resp = await fetch(`${server.url}my.jpg`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('image/jpeg');
+  });
+  it('should detect mime from plain Blob - GET', async () => {
+    app.get(`/my.png`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.png.data`
+      );
+      const buffer = await file.bytes();
+      const blob = new Blob([buffer], { type: 'image/png' });
+      return c.file(blob);
+    });
+    const resp = await fetch(`${server.url}my.png`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('image/png');
+  });
+  it('should detect mime from plain Blob - HEAD', async () => {
+    app.headGet(`/favicon.ico`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.ico.data`
+      );
+      const buffer = await file.bytes();
+      const blob = new Blob([buffer], { type: 'image/x-icon' });
+      return c.file(blob);
+    });
+    const resp = await fetch(`${server.url}favicon.ico`, { method: 'HEAD' });
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('image/x-icon');
+  });
+  it('should return 404 on bad input', async () => {
+    app.headGet(`/video.mov`, async c => {
+      // @ts-expect-error
+      return c.file(['foo']);
+    });
+    const resp = await fetch(`${server.url}video.mov`);
+    expect(resp.status).toBe(404);
+    expect(await resp.text()).toContain('File not found');
+  });
+  it('should allow disposition of attachment for Uint8Array', async () => {
+    app.get(`/font.woff2`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.woff2.data`
+      );
+      const buffer = await file.bytes();
+      return c.file(buffer, { disposition: 'attachment' });
+    });
+    const resp = await fetch(`${server.url}font.woff2`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('font/woff2');
+    expect(resp.headers.get('Content-Disposition')).toBe('attachment');
+  });
+  it('should allow disposition of attachment for Bunfile', async () => {
+    app.get(`/video.webm`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture-null.webm.data`
+      );
+      return c.file(file, { disposition: 'attachment' });
+    });
+    const resp = await fetch(`${server.url}video.webm`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('video/webm');
+    expect(resp.headers.get('Content-Disposition')).toBe(
+      'attachment; filename="fixture-null.webm.data"'
+    );
+  });
+  it('should allow disposition of form-data for Uint8Array', async () => {
+    app.get(`/my.tiff`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture-bali.tif`
+      );
+      const buffer = await file.bytes();
+      return c.file(buffer, { disposition: 'form-data' });
+    });
+    const resp = await fetch(`${server.url}my.tiff`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('image/tiff');
+    expect(resp.headers.get('Content-Disposition')).toBe('form-data');
+  });
+  it('should allow overriding content-type', async () => {
+    app.get(`/podcast.ogg`, async c => {
+      const file = Bun.file(
+        `${import.meta.dir}/../../../testFixtures/file-type/fixture.ogg.data`
+      );
+      const buffer = await file.bytes();
+      return c.file(buffer, { headers: { 'Content-Type': 'audio/ogg' } });
+    });
+    const resp = await fetch(`${server.url}podcast.ogg`);
+    expect(resp.status).toBe(200);
+    expect(resp.headers.get('Content-Type')).toInclude('audio/ogg');
+  });
 });