+++ packages/bunshine/src/responseFactories/json/json.spec.ts
@@ -8,7 +8,7 @@ describe('c.json()', () => {
   let server: Server;
   beforeEach(() => {
     app = new HttpRouter();
-    server = app.listen({ port: port++ });
+    server = app.listen({ port: 0 });
   });
   afterEach(() => {
     server.stop(true);