+++ packages/bunshine/src/responseFactories/sse/sse.spec.ts
@@ -16,7 +16,7 @@ describe('sse', () => {
   let server: Server;
   beforeEach(() => {
     app = new HttpRouter();
-    server = app.listen({ port: port++ });
+    server = app.listen({ port: 0 });
   });
   afterEach(() => {
     server.stop(true);
@@ -75,7 +75,7 @@ describe('sse', () => {
   it('should handle unnamed data', async () => {
     const events = await sseTest({
       event: 'message',
-      port: port++,
+      port: 0,
       payloads: [['Hello'], ['World']],
     });
     expect(events.length).toBe(2);
@@ -85,7 +85,7 @@ describe('sse', () => {
   it('should send last event id and origin', async () => {
     const events = await sseTest({
       event: 'myEvent',
-      port: port++,
+      port: 0,
       payloads: [
         ['myEvent', 'hi1', 'id1'],
         ['myEvent', 'hi2', 'id2'],
@@ -102,7 +102,7 @@ describe('sse', () => {
   it('should JSON encode data if needed', async () => {
     const events = await sseTest({
       event: 'myEvent',
-      port: port++,
+      port: 0,
       payloads: [['myEvent', { name: 'Bob' }]],
     });
     expect(events.length).toBe(1);
@@ -112,7 +112,7 @@ describe('sse', () => {
     spyOn(console, 'warn').mockImplementation(() => {});
     await sseTest({
       event: 'myEvent',
-      port: port++,
+      port: 0,
       payloads: [['myEvent', { name: 'Bob' }]],
       headers: {
         'Content-Type': 'text/plain',
@@ -128,7 +128,7 @@ describe('sse', () => {
     spyOn(console, 'warn').mockImplementation(() => {});
     await sseTest({
       event: 'myEvent',
-      port: port++,
+      port: 0,
       payloads: [['myEvent', { name: 'Bob' }]],
       headers: {
         'Content-Type': 'text/event-stream',